Revert "Update `snap_notify` to allow JSX content (#2706)"
This reverts commit 803d31440c83fd4a4cb45709c35e536107e6e2f2.
