+++ packages/snaps-sdk/src/jsx/validation.ts
@@ -618,28 +618,6 @@ export const SectionStruct: Describe<SectionElement> = element('Section', {
   ),
 });
 
-export const NotificationChildrenStruct = children(
-  // eslint-disable-next-line @typescript-eslint/no-use-before-define
-  [lazy(() => NotificationComponentsStruct)],
-) as unknown as Struct<SnapsChildren<GenericSnapElement>, null>;
-
-/**
- * A struct for the {@link BoxElement} type but with a limited subset of children.
- */
-export const NotificationStruct: Describe<BoxElement> = element('Box', {
-  children: NotificationChildrenStruct,
-  direction: optional(nullUnion([literal('horizontal'), literal('vertical')])),
-  alignment: optional(
-    nullUnion([
-      literal('start'),
-      literal('center'),
-      literal('end'),
-      literal('space-between'),
-      literal('space-around'),
-    ]),
-  ),
-});
-
 /**
  * A subset of JSX elements that are allowed as children of the Footer component.
  * This set should include a single button or a tuple of two buttons.
@@ -826,21 +804,6 @@ export const BoxChildStruct = typedUnion([
   SelectorStruct,
   SectionStruct,
   AvatarStruct,
-]);
-
-export const NotificationComponentsStruct = typedUnion([
-  CopyableStruct,
-  NotificationStruct,
-  BoldStruct,
-  ItalicStruct,
-  AddressStruct,
-  DividerStruct,
-  RowStruct,
-  TextStruct,
-  TooltipStruct,
-  IconStruct,
-  ImageStruct,
-  LinkStruct,
 ]);
 
 /**
