+++ packages/snaps-sdk/src/types/methods/notify.ts
@@ -1,19 +1,4 @@
-import { type EnumToUnion } from '../../internals';
-import type {
-  AddressElement,
-  BoldElement,
-  BoxElement,
-  CopyableElement,
-  DividerElement,
-  IconElement,
-  ImageElement,
-  ItalicElement,
-  LinkElement,
-  RowElement,
-  TextElement,
-  TooltipElement,
-  ValueElement,
-} from '../../jsx';
+import type { EnumToUnion } from '../../internals';
 
 /**
  * The types of notifications that can be displayed.
@@ -32,28 +17,10 @@ export enum NotificationType {
  * @property type - The type of notification to display.
  * @property message - The message to display in the notification.
  */
-export type NotifyParams =
-  | {
-      type: EnumToUnion<NotificationType.Native>;
-      message: string;
-    }
-  | {
-      type: EnumToUnion<NotificationType.InApp>;
-      message: string;
-    }
-  | {
-      type: EnumToUnion<NotificationType.InApp>;
-      message: string;
-      content: NotificationComponent;
-      title: string;
-    }
-  | {
-      type: EnumToUnion<NotificationType.InApp>;
-      message: string;
-      content: NotificationComponent;
-      title: string;
-      footerLink: { href: string; text: string };
-    };
+export type NotifyParams = {
+  type: EnumToUnion<NotificationType>;
+  message: string;
+};
 
 /**
  * The result returned by the `snap_notify` method.
@@ -61,18 +28,3 @@ export type NotifyParams =
  * This method does not return anything.
  */
 export type NotifyResult = null;
-
-export type NotificationComponent =
-  | BoxElement
-  | CopyableElement
-  | BoldElement
-  | ItalicElement
-  | AddressElement
-  | DividerElement
-  | ValueElement
-  | RowElement
-  | TextElement
-  | TooltipElement
-  | IconElement
-  | ImageElement
-  | LinkElement;
